+++ include/H2DE/H2DE.h
@@ -1,5 +1,5 @@
 /**
- * Version: 2.0.3
+ * Version: 2.0.4
  */
 
 #include "H2DE_engine.h"