+++ src/H2DE_circle.cpp
@@ -1,5 +1,0 @@
-#include "H2DE_renderer.h"
-
-void H2DE_Renderer::renderCircle(H2DE_GraphicObject* object) {
-
-}